refactor: month-bar Home에서 관리하게 이동
