+++ src/pages/Home.vue
@@ -1,22 +1,35 @@
 <template>
     <div class="outer">
         <div class="main-content">
-            <MonthlyExpenditure :recordList="recordList" @updateRecords="updateMonthlyRecords" />
-            <CategoryPieChart :records="monthlyRecords" />
-            <ToRegisterButton/>
+            <div class="month-bar">
+                <button @click="prevMonth" class="btn-icon"><i class="bi bi-caret-left-fill"></i></button>
+                <span class="currentMonth">{{ formattedDate }}</span>
+                <button @click="nextMonth" class="btn-icon"><i class="bi bi-caret-right-fill"></i></button>
+            </div>
+            <MonthlyExpenditure :recordList="recordList" 
+            :currentDate="currentDate"
+            :formattedDate="formattedDate"
+            @updateRecords="updateMonthlyRecords" />
+            <div class="visual-datas">
+                <CategoryPieChart :records="monthlyRecords" />
+                <YearlyProfitBarChart :recordList="recordList" @updateYear="updateCurrentYear" />
+            </div>
+            <ToRegisterButton />
         </div>
     </div>
 </template>
 
 <script setup>
-import { ref, onMounted } from 'vue'
+import { ref, onMounted, computed } from 'vue'
 import axios from "axios"
 import MonthlyExpenditure from '@/components/MonthlyExpenditure.vue'
 import CategoryPieChart from '@/components/CategoryPieChart.vue'
+import YearlyProfitBarChart from '@/components/YearlyProfitBarChart.vue'
 import ToRegisterButton from '@/components/ToRegisterButton.vue'
 
 const recordList = ref([])
 const monthlyRecords = ref([])
+const currentYear = ref(0)
 
 const getEveryRecords = async () => {
     const url = "http://localhost:3000/data"
@@ -28,9 +41,32 @@ const getEveryRecords = async () => {
     }
 }
 
+const currentDate = ref(new Date())
+
+const formattedDate = computed(() => {
+    return `${currentDate.value.getFullYear()}년 ${currentDate.value.getMonth() + 1}월`
+})
+
+
+const prevMonth = () => {
+    const date = new Date(currentDate.value)
+    date.setMonth(date.getMonth() - 1)
+    currentDate.value = date
+}
+
+const nextMonth = () => {
+    const date = new Date(currentDate.value)
+    date.setMonth(date.getMonth() + 1)
+    currentDate.value = date
+}
+
 // MonthlyExpenditure에서 추출한 해당 월 정보를 records에 업데이트, CategoryPieChart에서 사용하도록 함
 const updateMonthlyRecords = (records) => {
     monthlyRecords.value = records;
+}
+
+const updateCurrentYear = (year) => {
+    currentYear.value = year;
 }
 
 onMounted(() => {
@@ -39,11 +75,34 @@ onMounted(() => {
 </script>
 
 <style scoped>
-.outer{
+.outer {
     background-color: #D8EFD3;
-    border :2vh solid #D8EFD3;
-    min-height:98vh;
+    border: 2vh solid #D8EFD3;
+    min-height: 98vh;
     height: max-content;
+}
+
+.month-bar {
+    display: flex;
+    align-items: left;
+    justify-content: left;
+    gap: 10px;
+}
+
+.btn-icon {
+    display: inline-flex;
+    align-items: center;
+    justify-content: center;
+    background-color: transparent;
+    border: none;
+    color: black;
+    font-size: 18px;
+    cursor: pointer;
+}
+
+.currentMonth {
+    font-size: 18px;
+    font-weight: bold;
 }
 
 .main-content {
@@ -53,5 +112,11 @@ onMounted(() => {
     width: 80%;
     max-width: 1200px;
     border-radius: 8px;
+}
+
+.visual-datas {
+    display: flex;
+    justify-content: space-around;
+    gap: 40px;
 }
 </style>
